Convert saas/ folder to conform to eslint-plugin-react-refresh
