+++ frontend/src/core/auth/useAuth.ts
@@ -1,0 +1,1 @@
+export { useAuth } from '@app/auth/UseSession'
