+++ frontend/src/saas/components/auth/GuestUserBanner.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import CloseIcon from '@mui/icons-material/Close'
 import PersonAddIcon from '@mui/icons-material/PersonAdd'
-import { useAuth } from '@app/auth/UseSession'
+import { useAuth } from '@app/auth/useAuth'
 import { isUserAnonymous } from '@app/auth/supabase'
 import { withBasePath } from '@app/constants/app'
 import '@app/components/auth/GuestUserBanner.css'
