+++ frontend/src/saas/components/onboarding/useSaasOnboardingState.ts
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useMemo, useState, useRef } from 'react';
-import { useAuth } from '@app/auth/UseSession';
+import { useAuth } from '@app/auth/useAuth';
 import { useOs } from '@app/hooks/useOs';
 import {
   SLIDE_DEFINITIONS,
