+++ frontend/src/saas/hooks/useConfigButtonIcon.tsx
@@ -1,5 +1,5 @@
 import { Avatar } from '@mantine/core';
-import { useAuth } from '@app/auth/UseSession';
+import { useAuth } from '@app/auth/useAuth';
 
 export function useConfigButtonIcon(): React.ReactNode {
   const { profilePictureUrl } = useAuth();
