+++ frontend/src/saas/routes/Landing.tsx
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react'
 import { Navigate, useLocation } from 'react-router-dom'
-import { useAuth } from '@app/auth/UseSession'
+import { useAuth } from '@app/auth/useAuth'
 import { useAutoAnonymousAuth } from '@app/hooks/useAutoAnonymousAuth'
 import { isToolRoute } from '@app/utils/pathUtils'
 import HomePage from '@app/pages/HomePage'
